requirements.txt
datasets>=4.4.0
openai>=2.8.0
python-dotenv>=1.2.0
tiktoken>=0.8.0
typer>=0.12.5

